Moves utilities only used by `constant.py` out of `utils.py`, removes redundant dtype check in gather
